New UI implemented.
Code is _really_ bad, but we are racing against the clock so I just hacked things in.  At the least, need to leverage stylesheets _way_ more rather than inline styles.

Info boxes are gone, unfortunately.  Getting them in all of the right places is going to take more time than we have and we need to launch _something_ when Maker launches.