Increases min-height of Uniswap iframe to remove scrollbars.
I think Uniswap changed something on their end, as we now have scrollbars.  This fix just increases the min height to get rid of them.

Unfortunately, there is no way to auto-size an iframe correctly without the UI helping out, and in this case the Uniswap UI doesn't help out.

I probably won't push this change, as we don't have CI/CD and it is a bit of a pain to go through all of the motions of pushing, but I wanted to get it included in case we push for some other reason (or someone convinces me to push it).